略微修改代码
