+++ COG/Plugin/Loader/Controller/ClassType/Classes/Listener/PluginListener.cs
@@ -1,0 +1,6 @@
+﻿namespace COG.Plugin.Loader.Controller.ClassType.Classes.Listener;
+
+public interface IPluginListener
+{
+    
+}
